Convert h2 to h1
From a semantical point of view, because "My Todo List" is the "title"
of all the page content, it should be an h1. I should not reason about
how to style the page when defining the structure! 1) First, define the
page structure, and only 2) think about how things are formatted. Not
the other way round :-) (Beginner's mistake ;) )